+++ src/main/java/roomit/main/domain/chat/dto/ChatMessageResponse.java
@@ -1,4 +1,4 @@
-package roomit.web1_2_bumblebee_be.domain.chat.dto;
+package roomit.main.domain.chat.dto;
 
 import java.time.LocalDateTime;
