--- src/main/java/roomit/main/domain/chat/test.java
@@ -1,4 +1,0 @@
-package roomit.main.domain.chat;
-
-public class test {
-}
