+++ src/main/java/roomit/main/domain/chatroom/repositoroy/ChatRoomRepository.java
@@ -1,8 +1,8 @@
-package roomit.web1_2_bumblebee_be.domain.chatroom.repositoroy;
+package roomit.main.domain.chatroom.repositoroy;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import roomit.web1_2_bumblebee_be.domain.chatroom.entity.ChatRoom;
+import roomit.main.domain.chatroom.entity.ChatRoom;
 
 @Repository
 public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
