--- src/main/java/roomit/main/domain/chatroom/test.java
@@ -1,4 +1,0 @@
-package roomit.main.domain.chatroom;
-
-public class test {
-}
