+++ src/main/java/roomit/main/global/config/ScheduledTasks.java
@@ -1,9 +1,9 @@
-package roomit.web1_2_bumblebee_be.global.config;
+package roomit.main.global.config;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
-import roomit.web1_2_bumblebee_be.domain.chat.service.ChatService;
+import roomit.main.domain.chat.service.ChatService;
 
 @Component
 @RequiredArgsConstructor
